requirements.txt
requests==2.26.0
babel==2.9.0
django==3.1.7
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
django-oauth-toolkit==1.3.3
django-filter==2.4.0
django-unixtimestampfield==0.4.0
django-db-connection-pool[oracle,postgresql]==1.0.7
psycopg2-binary==2.8.6
markdown==3.3.4
celery[redis]==5.0.5
django-celery-beat==2.2.0
sentry-sdk==1.0.0
psutil==5.8.0
xlrd==1.2.0
pycryptodome==3.10.1
gevent==21.1.2
redis==3.5.3
python-decouple==3.4
pylint-django==2.4.4
django-cors-headers==3.7.0
oauth2client==4.1.3
gunicorn==20.1.0
python-dotenv==0.17.0
